Final found the shibboleth affiliation
Looks like it's going to be a nightmare to get Jetty to read the
shibboleth attribute as Env. variables (Which is the recommended way).
So I opted for a simple way i.e. putting them in the header, however this will require
updating the httpd shib.conf file with the option:
`ShibUseHeaders On`