Fix KeepsafeEntitlementRegistryClient.getPublisher()
The changes made to findMatchingEntitlement() were causing getPublisher()
to fail as it's also using the function but for a different purpose.

I've re-wrote findMatchingEntitlement() to match it's real use i.e.
search a matching entitlement for a given search term.

So intead of only querying for affiliation you can now use the function
to query the ER on any attributes (i.e. institution, publisher, etc) as
long as it matches the parameters expected by the ER.